refector: arrange module and test
