+++ src/index.test.js
@@ -1,18 +1,18 @@
 const sps = require('./index');
-const cmdCtrl = require('./command/ctrl');
+const proc = require('./proc');
 const config = require('./config');
 
 describe('main', () => {
   const key = 'key_1';
 
-  test('Run command by key', async () => {
+  test('Run proccess by key', async () => {
     config.read = jest.fn().mockResolvedValue({});
-    cmdCtrl.runByKey = jest.fn().mockResolvedValue();
+    proc.switchByKey = jest.fn().mockResolvedValue();
 
     const ret = await sps.main(key);
 
     expect(config.read).toHaveBeenCalledTimes(1);
-    expect(cmdCtrl.runByKey).toHaveBeenCalledTimes(1);
+    expect(proc.switchByKey).toHaveBeenCalledTimes(1);
     expect(ret).toBeUndefined();
   });
 });
